requirements.txt
flask
streamlit
pytest
langchain
langchain_ollama
selenium
beautifulsoup4
lxml
html5lib
python-dotenv
twilio
sqlite3